requirements.txt
quart
quart-cors
PyYAML